move scripts related to robody to RobodyControl
