+++ Assets/Scripts/InputDevices/SenseGlove/Prefabs/SG_RobotHand Right.prefab
