+++ Assets/Scripts/Training/ArmLengthCalibration/ArmLength.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using InputDevices;
 using UnityEngine;
-
+using RobodyControl;
 
 namespace Training.Calibration.ArmLength
 {
